+++ python/cugraph/graph/test_graph.py
@@ -14,9 +14,18 @@
 import numpy as np
 import pytest
 from scipy.io import mmread
-
 import cugraph
 import cudf
+
+# Temporarily suppress warnings till networkX fixes deprecation warnings
+# (Using or importing the ABCs from 'collections' instead of from
+# 'collections.abc' is deprecated, and in 3.8 it will stop working) for
+# python 3.7.  Also, this import networkx needs to be relocated in the
+# third-party group once this gets fixed.
+import warnings
+with warnings.catch_warnings():
+    warnings.filterwarnings("ignore", category=DeprecationWarning)
+    import networkx as nx
 
 
 def read_mtx_file(mm_file):
@@ -224,3 +233,38 @@ def test_two_hop_neighbors(graph_file):
     M = M.tocsr()
     find_two_paths(df, M)
     check_all_two_hops(df, M)
+
+
+@pytest.mark.parametrize('graph_file', DATASETS)
+def test_degree_functionality(graph_file):
+    M = read_mtx_file(graph_file)
+    sources = cudf.Series(M.row)
+    destinations = cudf.Series(M.col)
+    values = cudf.Series(M.data)
+
+    G = cugraph.Graph()
+    G.add_edge_list(sources, destinations, values)
+
+    Gnx = nx.DiGraph(M)
+
+    df_in_degree = G.in_degree()
+    df_out_degree = G.out_degree()
+    df_degree = G.degree()
+
+    nx_in_degree = Gnx.in_degree()
+    nx_out_degree = Gnx.out_degree()
+    nx_degree = Gnx.degree()
+
+    err_in_degree = 0
+    err_out_degree = 0
+    err_degree = 0
+    for i in range(len(df_degree)):
+        if(df_in_degree['degree'][i] != nx_in_degree[i]):
+            err_in_degree = err_in_degree + 1
+        if(df_out_degree['degree'][i] != nx_out_degree[i]):
+            err_out_degree = err_out_degree + 1
+        if(df_degree['degree'][i] != nx_degree[i]):
+            err_degree = err_degree + 1
+    assert err_in_degree == 0
+    assert err_out_degree == 0
+    assert err_degree == 0